refactor: Refactor slightly explorerkit-server
