+++ packages/explorerkit-server/src/server.ts
@@ -1,127 +1,33 @@
-import {
-  AccountParserInterface,
-  ErrorParserInterface,
-  InstructionParserInterface,
-  ParserType,
-  SolanaFMParser,
-} from "@solanafm/explorer-kit";
-import { getProgramIdl } from "@solanafm/explorer-kit-idls";
+import { AccountParserInterface, ParserType, SolanaFMParser } from "@solanafm/explorer-kit";
 import bodyParser from "body-parser";
-import { Buffer } from "buffer";
 import express, { Express, NextFunction, Request, Response } from "express";
-import NodeCache from "node-cache";
-import { collectDefaultMetrics,Gauge, Histogram, Registry } from "prom-client";
+import { collectDefaultMetrics, Histogram, Registry } from "prom-client";
 import { z } from "zod";
 
-interface Account {
-  ownerProgram: string;
-  data: string;
-}
+import { decodeProgramError } from "./decoders/errors";
+import { decodeInstruction, getProgramIds } from "./decoders/instructions";
+import { loadAllIdls, parsersCache } from "./parsers-cache";
+import { Account, DecodedAccount, TopLevelInstruction } from "./types";
+import { isValidBase58, isValidBase64 } from "./utils/validation";
 
 interface DecodeAccountsRequestBody {
   accounts: Account[];
-}
-
-interface DecodedAccount {
-  decodedData: DecodedAccountData | null;
-}
-
-interface DecodedAccountData {
-  owner: string;
-  name: string;
-  data: any;
 }
 
 interface DecodeTransactionsRequestBody {
   instructionsPerTransaction: (TopLevelInstruction[] | null)[];
 }
 
-interface TopLevelInstruction {
-  topLevelInstruction: Instruction;
-  flattenedInnerInstructions: Instruction[];
-}
-
-interface Instruction {
-  programId: string;
-  encodedData: string;
-  decodedData?: any | null;
-  name?: string | null;
-  accountKeys: string[];
-}
-
-interface ProgramError {
-  programId: string;
-  errorCode?: number | null;
-  decodedData?: any;
-  name?: string | null;
-}
-
 const register = new Registry();
 collectDefaultMetrics({ register });
 
-// Define custom metrics for NodeCache statistics
-const hitsGauge = new Gauge({
-  name: "nodecache_hits_total",
-  help: "Total number of cache hits",
-  registers: [register],
-});
-const missesGauge = new Gauge({
-  name: "nodecache_misses_total",
-  help: "Total number of cache misses",
-  registers: [register],
-});
-const keysGauge = new Gauge({
-  name: "nodecache_keys_total",
-  help: "Total number of keys in the cache",
-  registers: [register],
-});
-const ksizeGauge = new Gauge({
-  name: "nodecache_ksize_bytes",
-  help: "Total key size in bytes",
-  registers: [register],
-});
-const vsizeGauge = new Gauge({
-  name: "nodecache_vsize_bytes",
-  help: "Total value size in bytes",
-  registers: [register],
-});
 const httpRequestDurationMicroseconds = new Histogram({
-  name: 'http_request_duration_ms',
-  help: 'Duration of HTTP requests in ms',
-  labelNames: ['method', 'route', 'code'],
+  name: "http_request_duration_ms",
+  help: "Duration of HTTP requests in ms",
+  labelNames: ["method", "route", "code"],
   buckets: [0.1, 5, 15, 50, 100, 300, 500, 1000],
   registers: [register],
 });
-
-// Cache that evicts anything unused in the last 30mins
-let thirty_mins_in_seconds = 1800;
-const parsersCache = new NodeCache({ stdTTL: thirty_mins_in_seconds, checkperiod: 120 });
-
-// Update cache statistics in the metrics
-setInterval(() => {
-  const stats = parsersCache.getStats();
-  hitsGauge.set(stats.hits);
-  missesGauge.set(stats.misses);
-  keysGauge.set(stats.keys);
-  ksizeGauge.set(stats.ksize);
-  vsizeGauge.set(stats.vsize);
-}, 30000); // Update every 30sec
-
-// NOTE(fabio): If we cannot find a parser for a programId, we insert null into the cache
-// We want to periodically evict these null entries so that we retry fetching them periodically
-// in case they are now available
-function evictNullEntries(cache: NodeCache) {
-  const keys = cache.keys();
-  for (const key of keys) {
-    const value = cache.get(key);
-    if (value === null) {
-      cache.del(key);
-    }
-  }
-}
-
-const seventy_mins_in_miliseconds = 4200000;
-setInterval(evictNullEntries.bind(null, parsersCache), seventy_mins_in_miliseconds);
 
 const app: Express = express();
 app.use(bodyParser.json({ limit: "50mb" }));
@@ -129,13 +35,12 @@ app.use(bodyParser.json({ limit: "50mb" }));
 const responseDurationMiddleware = (req: Request, res: Response, next: NextFunction): void => {
   const start = process.hrtime();
 
-  res.on('finish', () => { // Event listener for when the response has been sent
+  res.on("finish", () => {
+    // Event listener for when the response has been sent
     const diff = process.hrtime(start);
     const responseTimeInMs = diff[0] * 1e3 + diff[1] * 1e-6; // Convert to milliseconds
 
-    httpRequestDurationMicroseconds
-      .labels(req.method, req.path, res.statusCode.toString())
-      .observe(responseTimeInMs);
+    httpRequestDurationMicroseconds.labels(req.method, req.path, res.statusCode.toString()).observe(responseTimeInMs);
   });
 
   next();
@@ -160,26 +65,27 @@ app.get("/metrics", async (_req: Request, res: Response) => {
 });
 
 const decodeAccountsSchema = z.object({
-  accounts: z.array(z.object({
-    ownerProgram: z.string().refine(isValidBase58, {
-      message: "account.ownerProgram is not a valid base58 string",
-
-    }),
-    data: z.string().refine(isValidBase64, {
-      message: "account.data is not a valid base64 string",
-    }),
-  })),
+  accounts: z.array(
+    z.object({
+      ownerProgram: z.string().refine(isValidBase58, {
+        message: "account.ownerProgram is not a valid base58 string",
+      }),
+      data: z.string().refine(isValidBase64, {
+        message: "account.data is not a valid base64 string",
+      }),
+    })
+  ),
 });
 
 // Endpoint to decode accounts data
 app.post("/decode/accounts", responseDurationMiddleware, async (req: Request, res: Response) => {
-  const {data, error} = decodeAccountsSchema.safeParse(req.body);
+  const { data, error } = decodeAccountsSchema.safeParse(req.body);
 
   if (error) {
     return res.status(400).json({ error: "Invalid request body", errors: error.errors });
   }
 
-  const {accounts} = data satisfies DecodeAccountsRequestBody;
+  const { accounts } = data satisfies DecodeAccountsRequestBody;
 
   let allProgramIds = [];
   for (let account of accounts) {
@@ -215,45 +121,61 @@ app.post("/decode/accounts", responseDurationMiddleware, async (req: Request, re
 });
 
 const decodeErrorsSchema = z.object({
-  errors: z.array(z.object({
-    programId: z.string(),
-    errorCode: z.coerce.number().nullable().optional(),
-  }).nullable()),
+  errors: z.array(
+    z
+      .object({
+        programId: z.string(),
+        errorCode: z.coerce.number().nullable().optional(),
+      })
+      .nullable()
+  ),
 });
 
 // Endpoint to decode program errors
 app.post("/decode/errors", responseDurationMiddleware, async (req: Request, res: Response) => {
-  const {data, error} = decodeErrorsSchema.safeParse(req.body);
+  const { data, error } = decodeErrorsSchema.safeParse(req.body);
 
   if (error) {
     return res.status(400).json({ error: "Invalid request body", errors: error.errors });
   }
 
-  const programIdsWithFailure = data.errors.filter(err => err?.errorCode).map((err) => err?.programId).map(v => v!);
+  const programIdsWithFailure = data.errors
+    .filter((err) => err?.errorCode)
+    .map((err) => err?.programId)
+    .map((v) => v!);
+
   await loadAllIdls(programIdsWithFailure);
-  const decodedErrors = data.errors.map((error) => error && decodeProgramError(error))
+  const decodedErrors = data.errors.map((error) => error && decodeProgramError(error));
 
   return res.status(200).json({ decodedErrors });
-})
+});
 
 const decodeInstructionsSchema = z.object({
-  instructionsPerTransaction: z.array(z.array(z.object({
-    topLevelInstruction: z.object({
-      programId: z.string(),
-      encodedData: z.string(),
-      accountKeys: z.array(z.string()),
-    }),
-    flattenedInnerInstructions: z.array(z.object({
-      programId: z.string(),
-      encodedData: z.string(),
-      accountKeys: z.array(z.string()),
-    })),
-  })).nullable()),
+  instructionsPerTransaction: z.array(
+    z
+      .array(
+        z.object({
+          topLevelInstruction: z.object({
+            programId: z.string(),
+            encodedData: z.string(),
+            accountKeys: z.array(z.string()),
+          }),
+          flattenedInnerInstructions: z.array(
+            z.object({
+              programId: z.string(),
+              encodedData: z.string(),
+              accountKeys: z.array(z.string()),
+            })
+          ),
+        })
+      )
+      .nullable()
+  ),
 });
 
 // Endpoint to decode instructions for a list of transactions
 app.post("/decode/instructions", responseDurationMiddleware, async (req: Request, res: Response) => {
-  const {data, error} = decodeInstructionsSchema.safeParse(req.body);
+  const { data, error } = decodeInstructionsSchema.safeParse(req.body);
 
   if (error) {
     return res.status(400).json({ error: "Invalid request body", errors: error.errors });
@@ -293,165 +215,5 @@ app.post("/decode/instructions", responseDurationMiddleware, async (req: Request
     return res.status(500).json({ error: e.message });
   }
 });
-
-async function loadAllIdls(programIds: string[]) {
-  let getProgramIdlFutures = [];
-  let programsWithMissingIdls = [];
-  for (var programId of programIds) {
-    let instructionParser = parsersCache.get(programId) as InstructionParserInterface;
-    // If we already seen the programId before, skip
-    if (instructionParser !== undefined) {
-      continue;
-    }
-
-    // We haven't seen the programId before, add call to fetch it's IDL
-    getProgramIdlFutures.push(getProgramIdl(programId));
-    programsWithMissingIdls.push(programId);
-  }
-
-  let SFMIdlItems = await Promise.all(getProgramIdlFutures);
-  for (var i = 0; i < SFMIdlItems.length; i++) {
-    let SFMIdlItem = SFMIdlItems[i];
-    let programId = programsWithMissingIdls[i] as string;
-    if (SFMIdlItem != undefined) {
-      const parser = new SolanaFMParser(SFMIdlItem, programId);
-      parsersCache.set(programId, parser);
-    } else {
-      parsersCache.set(programId, null); // Insert into cache to avoid trying to fetch IDL again
-    }
-  }
-}
-
-async function decodeInstruction(instruction: Instruction): Promise<Instruction> {
-  const programId = instruction.programId.toString();
-  let parsedInstruction = {
-    programId: programId.toString(),
-    encodedData: instruction.encodedData,
-    name: null,
-    decodedData: null,
-    accountKeys: instruction.accountKeys,
-  };
-
-  let parser = parsersCache.get(programId) as SolanaFMParser;
-  if (parser == null) {
-    return parsedInstruction; // Short-circuit without decodedData since IDL is missing
-  }
-  let instructionParser = parser.createParser(ParserType.INSTRUCTION) as InstructionParserInterface;
-
-  // Parse the transaction
-  const decodedInstruction = instructionParser.parseInstructions(instruction.encodedData, instruction.accountKeys);
-  const decodedInstructionWithTypes = instructionParser.parseInstructions(
-    instruction.encodedData,
-    instruction.accountKeys,
-    true
-  );
-  const finalDecodedInstructionData = postProcessDecodedInstruction(
-    decodedInstruction?.data,
-    decodedInstructionWithTypes?.data
-  );
-  return {
-    programId: instruction.programId,
-    encodedData: instruction.encodedData,
-    name: decodedInstruction?.name || null,
-    decodedData: finalDecodedInstructionData || null,
-    accountKeys: instruction.accountKeys,
-  };
-}
-
-function decodeProgramError(programError: ProgramError): ProgramError {
-  if (!programError.errorCode) {
-    return programError
-  }
-
-  const programId = programError.programId;
-  const parser = parsersCache.get<SolanaFMParser>(programId);
-
-  if (!parser) {
-    return programError;
-  }
-
-  const hexErrorCode = `0x${programError.errorCode.toString(16)}`;
-  const errorParser = parser.createParser(ParserType.ERROR) as ErrorParserInterface;
-  const parsedError = errorParser.parseError(hexErrorCode);
-
-  if (!parsedError) {
-    return programError;
-  }
-
-  return {
-    ...programError,
-    decodedData: parsedError.data,
-    name: parsedError.name,
-  }
-}
-
-function postProcessDecodedInstruction(decodedInstructionData?: any, decodedInstructionDataWithTypes?: any): any {
-  if (!decodedInstructionData || !decodedInstructionDataWithTypes) {
-    return null;
-  }
-
-  Object.keys(decodedInstructionDataWithTypes).forEach((key) => {
-    const property = decodedInstructionDataWithTypes[key];
-    // If [[u8, X]],  base64 encode it
-    if (
-      typeof property.type === "object" &&
-      typeof property.type.vec === "object" &&
-      Array.isArray(property.type.vec.array) &&
-      property.type.vec.array[0] === "u8"
-    ) {
-      const rawValue = decodedInstructionData[key];
-      if (rawValue && Array.isArray(rawValue) && Array.isArray(rawValue[0])) {
-        decodedInstructionData[key] = rawValue.map((arr: number[]) => Buffer.from(arr).toString("base64"));
-      }
-    }
-    // If [u8, X], base64 encode it
-    if (typeof property.type === "object" && Array.isArray(property.type.array) && property.type.array[0] === "u8") {
-      // Base64 encode the byte array from decodedInstructionData
-      const byteArray = decodedInstructionData[key];
-      if (byteArray && Array.isArray(byteArray)) {
-        decodedInstructionData[key] = Buffer.from(byteArray).toString("base64");
-      }
-    }
-    // If bytes, base64 encode it
-    if (property.type === "bytes") {
-      const bytes = decodedInstructionData[key];
-      if (bytes && bytes.constructor === Uint8Array) {
-        decodedInstructionData[key] = Buffer.from(bytes).toString("base64");
-      }
-    }
-  });
-
-  return decodedInstructionData;
-}
-
-function getProgramIds(instructionsPerTransaction: (TopLevelInstruction[] | null)[]): string[] {
-  let allProgramIds: string[] = [];
-  for (var transactionInstructions of instructionsPerTransaction) {
-    if (transactionInstructions === null) continue; // Skip nulls
-    for (var instruction of transactionInstructions) {
-      allProgramIds.push(instruction.topLevelInstruction.programId);
-      for (var inner_instruction of instruction.flattenedInnerInstructions) {
-        allProgramIds.push(inner_instruction.programId);
-      }
-    }
-  }
-  // Dedup programIds
-  let dedupedProgramIds = new Set(allProgramIds);
-  return Array.from(dedupedProgramIds);
-}
-
-function isValidBase58(str: string): boolean {
-  const base58Regex = /^[A-HJ-NP-Za-km-z1-9]+$/;
-  return base58Regex.test(str);
-}
-
-function isValidBase64(str: string): boolean {
-  try {
-    const base64Encoded = Buffer.from(str, "base64").toString("base64");
-    return str === base64Encoded;
-  } catch (e) {
-    return false;
-  }
-}
 
 export { app };
