+++ packages/explorerkit-server/src/utils/validation.ts
@@ -1,0 +1,15 @@
+import { Buffer } from "buffer";
+
+export function isValidBase58(str: string): boolean {
+  const base58Regex = /^[A-HJ-NP-Za-km-z1-9]+$/;
+  return base58Regex.test(str);
+}
+
+export function isValidBase64(str: string): boolean {
+  try {
+    const base64Encoded = Buffer.from(str, "base64").toString("base64");
+    return str === base64Encoded;
+  } catch (e) {
+    return false;
+  }
+}
